interop/python-runner/requirements.txt
atoma-sdk>=0.1.0
python-dotenv>=1.0.0
